# Changed registry's mutex to be RWMutex