For performance gains when Get and GetOrRegister are called frequently