@mihasya 2018-04-06
Sorry about the delay in replying.

Did this show up in profile? Have you considered keeping a local pointer to the metric around instead of calling a Getter each time? I'm always suspicious when the getters show up in a profile for this library. Most likely, restructuring the code will net you larger gains than even the patch. My apologies if this is something you've already investigated and did not find sufficient.

All that said, this change does seem correct. If you could add some benchmarks that illustrate the gains, as well as some test that would exercise this functionality for the race detector, I would merge it right away. It all "looks good," but as we saw in #232, nothing is quite as good as an actual test.