@marcusking01 2018-04-06
@mihasya In 90% of the cases we do hold a local pointer to a metric and use that reference, however in some cases where we have dynamic metric names we do use the GetOrRegister which would be optimized by using a RWMutex.  I'll work on getting a benchmark and test together, I actually worked with the developer on #232 so I understand the caution here.