@marcusking01 2018-04-06
Previous Code (Mutex)
`
    BenchmarkRegistry-8           	          5000000	       297 ns/op
    BenchmarkRegistryParallel-8   	20000000	       111 ns/op
    PASS
    ok  	_/Users/mking/workspace/github/go-metrics	10.468s
`

New Code (RWMutex)
` 
    BenchmarkRegistry-8           	          5000000	       303 ns/op
    BenchmarkRegistryParallel-8   	30000000	        50.5 ns/op
    PASS
    ok  	_/Users/mking/workspace/github/go-metrics	9.132s
`