@mihasya 2018-04-06
Ok this looks awesome! Just to keep things tidy, would you mind creating a new branch off of current `master` and cherry-picking your two commits onto it? That will remove the extra "merge" commit (this can be avoided by not mergin master into the branch, and instead doing a rebase -  `git checkout master && git pull origin master && git checkout mybranch && git rebase master`, then resolving conflicts on a per-commit basis. Likely this would have rebase with no conflicts, as the PR said it was clean to merge). Just keeps the commit history clean.